feat: ajout d'un contrôle sur les commandes requises (#3)
* feat: ajout d'un contrôle sur les commandes requises

* fix: prise en compte du contexte CI sur le contrôle des dépendances

+++ .bin/scripts/deploy-log-decrypt.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#    !/usr/bin/env bash
 
 set -euo pipefail
 
@@ -26,7 +26,9 @@ readonly PASSPHRASE="$ROOT_DIR/.bin/SEED_PASSPHRASE.txt"
 readonly VAULT_FILE="${ROOT_DIR}/.infra/vault/vault.yml"
 
 delete_cleartext() {
-  shred -f -n 10 -u "$PASSPHRASE"
+  if [ -f "$PASSPHRASE" ]; then
+    shred -f -n 10 -u "$PASSPHRASE"
+  fi
 }
 trap delete_cleartext EXIT
